+++ includes/Open62541_Server/NodeBuilder.hpp
@@ -16,40 +16,46 @@ class NodeBuilder {
   std::shared_ptr<Open62541Server> server_;
 
   std::pair<UA_StatusCode, UA_NodeId> addObjectNode(
-      Information_Model::NonemptyNamedElementPtr element,
+      const Information_Model::NonemptyNamedElementPtr& element,
       std::optional<UA_NodeId> parent_node_id = std::nullopt);
 
   UA_StatusCode addDeviceNodeElement(
-      Information_Model::NonemptyDeviceElementPtr device_element,
-      UA_NodeId parent_id);
+      const Information_Model::NonemptyDeviceElementPtr& device_element,
+      const UA_NodeId& parent_id);
 
   UA_StatusCode addGroupNode(
-      Information_Model::NonemptyNamedElementPtr meta_info,
-      Information_Model::NonemptyDeviceElementGroupPtr device_element_group,
-      UA_NodeId parent_id);
-
-  UA_StatusCode addFunctionNode(
-      Information_Model::DeviceElementPtr function, UA_NodeId parent_id);
+      const Information_Model::NonemptyNamedElementPtr& meta_info,
+      const Information_Model::NonemptyDeviceElementGroupPtr&
+          device_element_group,
+      const UA_NodeId& parent_id);
 
   UA_StatusCode addReadableNode(
-      Information_Model::NonemptyNamedElementPtr meta_info,
-      Information_Model::NonemptyMetricPtr metric, UA_NodeId parent_id);
+      const Information_Model::NonemptyNamedElementPtr& meta_info,
+      const Information_Model::NonemptyMetricPtr& metric,
+      const UA_NodeId& parent_id);
 
   UA_StatusCode addWritableNode(
-      Information_Model::NonemptyNamedElementPtr meta_info,
-      Information_Model::NonemptyWritableMetricPtr metric, UA_NodeId parent_id);
+      const Information_Model::NonemptyNamedElementPtr& meta_info,
+      const Information_Model::NonemptyWritableMetricPtr& metric,
+      const UA_NodeId& parent_id);
+
+  UA_StatusCode addFunctionNode(
+      const Information_Model::NonemptyNamedElementPtr& meta_info,
+      const Information_Model::NonemptyFunctionPtr& function,
+      const UA_NodeId& parent_id);
 
   template <class MetricType> // either Metric or WritableMetric
   UA_StatusCode setValue(UA_VariableAttributes& value_attribute,
-      Information_Model::NonemptyNamedElementPtr meta_info,
-      NonemptyPointer::NonemptyPtr<std::shared_ptr<MetricType>> metric,
-      std::string metric_type_description);
+      const Information_Model::NonemptyNamedElementPtr& meta_info,
+      const NonemptyPointer::NonemptyPtr<std::shared_ptr<MetricType>>& metric,
+      const std::string& metric_type_description);
 
 public:
-  NodeBuilder(std::shared_ptr<Open62541Server> server);
+  NodeBuilder(const std::shared_ptr<Open62541Server>& server);
   ~NodeBuilder();
 
-  UA_StatusCode addDeviceNode(Information_Model::NonemptyDevicePtr device);
+  UA_StatusCode addDeviceNode(
+      const Information_Model::NonemptyDevicePtr& device);
   UA_StatusCode deleteDeviceNode(const std::string& device_id);
   void cleanup();
 };