+++ includes/Open62541_Server/Open62541Server.hpp
@@ -27,7 +27,7 @@ class Open62541Server {
   std::unique_ptr<Historizer> historizer_;
 
   UA_StatusCode registerForHistorization(
-      UA_NodeId nodeId, const UA_DataType* type);
+      UA_NodeId node_id, const UA_DataType* type);
 #endif // UA_ENABLE_HISTORIZING
 
   void runnable();
@@ -50,8 +50,8 @@ public:
   bool start();
   bool stop();
   bool isRunning();
-  UA_UInt16 getServerNamespace();
-  const UA_Logger* getServerLogger();
+  UA_UInt16 getServerNamespace() const;
+  const UA_Logger* getServerLogger() const;
   UA_Server* getServer();
 };
 } // namespace open62541